faceEmotion/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
opencv-python-headless
pillow
numpy
python-multipart
